move api/system/time to robot-server. Tests will fail.
